Added seperate HW detection
